Solution

+++ src/routes/accounts.py
@@ -1,7 +1,7 @@
 from datetime import datetime, timezone
-from typing import cast
+from typing import cast, Annotated
 
-from fastapi import APIRouter, Depends, status, HTTPException
+from fastapi import APIRouter, Depends, status, HTTPException, BackgroundTasks
 from sqlalchemy import select, delete
 from sqlalchemy.exc import SQLAlchemyError
 from sqlalchemy.ext.asyncio import AsyncSession
@@ -35,6 +35,22 @@ from security.interfaces import JWTAuthManagerInterface
 
 router = APIRouter()
 
+ACTIVATION_LINK = "http://127.0.0.1/activate/"
+LOGIN_LINK = "http://127.0.0.1/accounts/login/"
+PASSWORD_RESET_LINK = "http://127.0.0.1/password-reset/request/"
+PASSWORD_RESET_COMPLETE_LINK = "http://127.0.0.1/reset-password/complete/"
+
+
+async def get_user_group(db: Annotated[AsyncSession, Depends(get_db)]) -> None:
+    stmt = select(UserGroupModel).where(UserGroupModel.name == UserGroupEnum.USER)
+    result = await db.execute(stmt)
+    user_group = result.scalars().first()
+    if not user_group:
+        raise HTTPException(
+            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
+            detail="Default user group not found."
+        )
+
 
 @router.post(
     "/register/",
@@ -67,7 +83,10 @@ router = APIRouter()
 )
 async def register_user(
         user_data: UserRegistrationRequestSchema,
-        db: AsyncSession = Depends(get_db),
+        user_group: Depends(get_user_group),
+        email: Annotated[EmailSenderInterface, Depends(get_accounts_email_notificator)],
+        db: Annotated[AsyncSession, Depends(get_db)],
+        background_tasks: BackgroundTasks
 ) -> UserRegistrationResponseSchema:
     """
     Endpoint for user registration.
@@ -79,6 +98,9 @@ async def register_user(
     Args:
         user_data (UserRegistrationRequestSchema): The registration details including email and password.
         db (AsyncSession): The asynchronous database session.
+        email (EmailSenderInterface): The email functions interface
+        user_group: The user group retrieve.
+        background_tasks: Fast api background task class.
 
     Returns:
         UserRegistrationResponseSchema: The newly created user's details.
@@ -97,15 +119,6 @@ async def register_user(
             detail=f"A user with this email {user_data.email} already exists."
         )
 
-    stmt = select(UserGroupModel).where(UserGroupModel.name == UserGroupEnum.USER)
-    result = await db.execute(stmt)
-    user_group = result.scalars().first()
-    if not user_group:
-        raise HTTPException(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            detail="Default user group not found."
-        )
-
     try:
         new_user = UserModel.create(
             email=str(user_data.email),
@@ -117,6 +130,12 @@ async def register_user(
 
         activation_token = ActivationTokenModel(user_id=new_user.id)
         db.add(activation_token)
+
+        background_tasks.add_task(
+            email.send_activation_email,
+            email=new_user.email,
+            activation_link=ACTIVATION_LINK + f"?activation_token={activation_token.token}"
+        )
 
         await db.commit()
         await db.refresh(new_user)
@@ -131,7 +150,7 @@ async def register_user(
 
 
 @router.post(
-    "/activate/",
+    "/activate/{activation_token}",
     response_model=MessageResponseSchema,
     summary="Activate User Account",
     description="Activate a user's account using their email and activation token.",
@@ -162,8 +181,11 @@ async def register_user(
     },
 )
 async def activate_account(
+        activation_token: str,
         activation_data: UserActivationRequestSchema,
-        db: AsyncSession = Depends(get_db),
+        email: Annotated[EmailSenderInterface, Depends(get_accounts_email_notificator)],
+        db: Annotated[AsyncSession, Depends(get_db)],
+        background_tasks: BackgroundTasks
 ) -> MessageResponseSchema:
     """
     Endpoint to activate a user's account.
@@ -175,8 +197,10 @@ async def activate_account(
 
     Args:
         activation_data (UserActivationRequestSchema): Contains the user's email and activation token.
+        activation_token: User activation token.
+        email (EmailSenderInterface): The email functions interface.
         db (AsyncSession): The asynchronous database session.
-
+        background_tasks: Fast api background task class.
     Returns:
         MessageResponseSchema: A response message confirming successful activation.
 
@@ -199,9 +223,15 @@ async def activate_account(
 
     now_utc = datetime.now(timezone.utc)
     if not token_record or cast(datetime, token_record.expires_at).replace(tzinfo=timezone.utc) < now_utc:
-        if token_record:
+        if token_record and token_record.token == activation_token:
             await db.delete(token_record)
             await db.commit()
+
+            background_tasks.add_task(
+                email.send_activation_complete_email,
+                email=activation_data.email,
+                login_link=LOGIN_LINK
+            )
         raise HTTPException(
             status_code=status.HTTP_400_BAD_REQUEST,
             detail="Invalid or expired activation token."
@@ -233,7 +263,9 @@ async def activate_account(
 )
 async def request_password_reset_token(
         data: PasswordResetRequestSchema,
-        db: AsyncSession = Depends(get_db),
+        email: Annotated[EmailSenderInterface, Depends(get_accounts_email_notificator)],
+        db: Annotated[AsyncSession, Depends(get_db)],
+        background_tasks: BackgroundTasks
 ) -> MessageResponseSchema:
     """
     Endpoint to request a password reset token.
@@ -243,7 +275,9 @@ async def request_password_reset_token(
 
     Args:
         data (PasswordResetRequestSchema): The request data containing the user's email.
+        email (EmailSenderInterface): The email functions interface
         db (AsyncSession): The asynchronous database session.
+        background_tasks: Fast api background task class.
 
     Returns:
         MessageResponseSchema: A success message indicating that instructions will be sent.
@@ -261,6 +295,12 @@ async def request_password_reset_token(
 
     reset_token = PasswordResetTokenModel(user_id=cast(int, user.id))
     db.add(reset_token)
+    background_tasks.add_task(
+        email.send_password_reset_email,
+        email=data.email,
+        reset_link=PASSWORD_RESET_COMPLETE_LINK + f"?reset_token={reset_token.token}"
+    )
+
     await db.commit()
 
     return MessageResponseSchema(
@@ -269,7 +309,7 @@ async def request_password_reset_token(
 
 
 @router.post(
-    "/reset-password/complete/",
+    "/reset-password/complete/{reset_token}",
     response_model=MessageResponseSchema,
     summary="Reset User Password",
     description="Reset a user's password if a valid token is provided.",
@@ -312,8 +352,11 @@ async def request_password_reset_token(
     },
 )
 async def reset_password(
+        reset_token: str,
         data: PasswordResetCompleteRequestSchema,
-        db: AsyncSession = Depends(get_db),
+        email: Annotated[EmailSenderInterface, Depends(get_accounts_email_notificator)],
+        db: Annotated[AsyncSession, Depends(get_db)],
+        background_tasks: BackgroundTasks
 ) -> MessageResponseSchema:
     """
     Endpoint for resetting a user's password.
@@ -322,9 +365,12 @@ async def reset_password(
     Deletes the token after a successful password reset.
 
     Args:
+        reset_token: User reset token.
         data (PasswordResetCompleteRequestSchema): The request data containing the user's email,
          token, and new password.
+        email (EmailSenderInterface): The email functions interface.
         db (AsyncSession): The asynchronous database session.
+        background_tasks: Fast api background task class.
 
     Returns:
         MessageResponseSchema: A response message indicating successful password reset.
@@ -348,9 +394,15 @@ async def reset_password(
     token_record = result.scalars().first()
 
     if not token_record or token_record.token != data.token:
-        if token_record:
+        if token_record and token_record.token == reset_token:
             await db.run_sync(lambda s: s.delete(token_record))
             await db.commit()
+            background_tasks.add_task(
+                email.send_password_reset_complete_email(
+                    email=user.email,
+                    login_link=LOGIN_LINK
+                )
+            )
         raise HTTPException(
             status_code=status.HTTP_400_BAD_REQUEST,
             detail="Invalid email or token."
@@ -420,7 +472,7 @@ async def reset_password(
 )
 async def login_user(
         login_data: UserLoginRequestSchema,
-        db: AsyncSession = Depends(get_db),
+        db: Annotated[AsyncSession, Depends(get_db)],
         settings: BaseAppSettings = Depends(get_settings),
         jwt_manager: JWTAuthManagerInterface = Depends(get_jwt_auth_manager),
 ) -> UserLoginResponseSchema:
@@ -527,7 +579,7 @@ async def login_user(
 )
 async def refresh_access_token(
         token_data: TokenRefreshRequestSchema,
-        db: AsyncSession = Depends(get_db),
+        db: Annotated[AsyncSession, Depends(get_db)],
         jwt_manager: JWTAuthManagerInterface = Depends(get_jwt_auth_manager),
 ) -> TokenRefreshResponseSchema:
     """
